fix(theme-shadcn): rewrite all components from document.createElement to JSX (#1639)
All theme-shadcn components used document.createElement() to create elements
imperatively. During SSR hydration, the Vertz runtime's __element(tag) claims
existing SSR nodes from the DOM, but document.createElement bypasses this
mechanism — creating new detached elements that don't match SSR nodes.

Converted all 13 remaining components to JSX (.ts → .tsx):
- card, input, label, badge, textarea, table, avatar, alert
- form-group, separator, skeleton, breadcrumb, pagination

JSX compiles to __element() calls which correctly participate in hydration.
Components with spread props (input, textarea) use applyProps() as a
workaround for #1633 (compiler JSX spread limitation).

For breadcrumb and pagination, list items are built in imperative loops
rather than .map() inside JSX, working around a compiler limitation where
__list() doesn't recursively transform JSX in callback bodies.

Fixes #1639
